common: move old bad blocks API to commons
