+++ src/test/util_badblock/util_badblock.c
@@ -11,7 +11,6 @@
 #include "out.h"
 #include "set.h"
 #include "badblocks.h"
-#include "os_badblock.h"
 #include "set_badblocks.h"
 #include "fault_injection.h"
 #include "file.h"
