+++ src/tools/pmempool/info.c
@@ -26,7 +26,6 @@
 #include "set.h"
 #include "file.h"
 #include "badblocks.h"
-#include "os_badblock.h"
 #include "set_badblocks.h"
 
 #define DEFAULT_CHUNK_TYPES\
